pyproject.toml
[project]
name = "segma"
version = "0.1.0"
description = "Minimal audio segmentation toolkit."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asteroid-filterbanks>=0.4.0",
    "dacite>=1.9.2",
    "interlap>=0.2.7",
    "lightning>=2.4.0",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "omegaconf>=2.3.0",
    "optuna>=4.2.1",
    "polars>=1.22.0",
    "pyyaml>=6.0.2",
    "scipy>=1.14.1",
    "soundfile>=0.12.1; sys_platform == 'darwin'",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "transformers>=4.47.1",
    "wandb>=0.18.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "notebook>=7.2.2",
    "basedpyright>=1.23.2",
]

[tool.ruff.lint]
ignore = ["E722"]

[tool.mypy]
exclude = ["scripts"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*sentry_sdk\\.Hub.*:DeprecationWarning",
]
